First round of upgrades to RF_sensor

+++ src/RF_sensor.h
@@ -9,7 +9,7 @@ typedef union {
         uint16_t forward;
         uint16_t reverse;
         double swr;
-        uint16_t period;
+        uint32_t period;
         uint16_t frequency;
     };
 } RF_power_s;
@@ -32,7 +32,7 @@ extern void SWR_threshold_increment(void);
 /* -------------------------------------------------------------------------- */
 
 // Frequency measurement
-extern uint16_t get_period(void);
+extern uint32_t get_period(void);
 
 // SWR measurement functions
 extern void SWR_measure(void);
